MojoHandle -> uintptr_t
MojoHandle was recently changed from 32 bits to 64 bits wide in order
to have parity with ipcz handles. ipcz handles however have since been
standardized as pointer-sized rather than fixed at 64 bits.

This updates MojoHandle to be an alias for uintptr_t, in order to
maintain parity with current ipcz, once again breaking the Mojo
ABI. This is fine since the Mojo ABI only has one known project
depending on its stability (Chrome OS IME) and it's already using
an alternative (strictly 32-bit) ABI.

Bug: 1299283
Fixed: 1322162
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3804110
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/main@{#1031404}

+++ ipc/ipc_mojo_param_traits.cc
@@ -39,7 +39,7 @@ bool ParamTraits<mojo::MessagePipeHandle>::Read(const base::Pickle* m,
 void ParamTraits<mojo::MessagePipeHandle>::Log(const param_type& p,
                                                std::string* l) {
   l->append("mojo::MessagePipeHandle(");
-  LogParam(p.value(), l);
+  LogParam(static_cast<uint64_t>(p.value()), l);
   l->append(")");
 }
 
@@ -89,7 +89,7 @@ bool ParamTraits<mojo::DataPipeConsumerHandle>::Read(const base::Pickle* m,
 void ParamTraits<mojo::DataPipeConsumerHandle>::Log(const param_type& p,
                                                     std::string* l) {
   l->append("mojo::DataPipeConsumerHandle(");
-  LogParam(p.value(), l);
+  LogParam(static_cast<uint64_t>(p.value()), l);
   l->append(")");
 }
